refactor(ui): collapse Detection + DebugConsole activities into bottom sheets (T8)
Port two compact Activities to BottomSheetDialogFragments now that the v2.0
shell launches Material sheets instead of separate screens.

Ported:
- detection.DetectionTestActivity  -> ui.DetectionTestBottomSheet
- debug.DebugConsoleActivity       -> ui.DebugConsoleBottomSheet

Manifest entries dropped:
- <activity .detection.DetectionTestActivity> (v3.0 T4.2)
- <activity .debug.DebugConsoleActivity>      (v3.0 T4.3)

Placeholder layouts hold every original view id so the ported Java keeps
working unchanged; T15 will rebuild the surfaces. Lifecycle adjustments:
view cleanup moved to onDestroyView, runOnUiThread routed through
requireActivity() with an isAdded()/isRemoving() guard, listener
add/remove kept on onStart/onStop.

MainActivity's detection button now shows DetectionTestBottomSheet via
showDetectionSheet(). DebugConsoleBottomSheet exposes a public static
show(FragmentManager) helper but remains UI-unreachable until T10 lands
the 7-tap unlock on the new AboutActivity (or T12 wires it from the
Settings sheet) — open concern carried from T7.
