+++ app/src/main/res/layout/bottom_sheet_debug_console.xml
@@ -1,13 +1,18 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.appcompat.widget.LinearLayoutCompat
     xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:id="@+id/debug_layout"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
+    android:layout_height="wrap_content"
+    android:background="?attr/colorSurface"
+    android:minHeight="320dp"
     android:orientation="vertical"
-    android:padding="16dp">
+    android:padding="@dimen/spacing_md">
+
+    <include layout="@layout/view_drag_handle" />
 
     <com.google.android.flexbox.FlexboxLayout
-        xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         app:flexWrap="wrap">
@@ -20,7 +25,7 @@
             android:text="@string/Debug_Export" />
 
         <Space
-            android:layout_width="8dp"
+            android:layout_width="@dimen/spacing_sm"
             android:layout_height="wrap_content" />
 
         <com.google.android.material.button.MaterialButton
@@ -31,7 +36,7 @@
             android:text="@string/Debug_Clear" />
 
         <Space
-            android:layout_width="8dp"
+            android:layout_width="@dimen/spacing_sm"
             android:layout_height="wrap_content" />
 
         <com.google.android.material.button.MaterialButton
@@ -45,6 +50,8 @@
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/log_list"
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_marginTop="12dp" />
+        android:layout_height="0dp"
+        android:layout_marginTop="@dimen/spacing_sm"
+        android:layout_weight="1" />
+
 </androidx.appcompat.widget.LinearLayoutCompat>
